requirements.txt
aiosmtplib==2.0.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==2.0.2
bcrypt==4.0.1
cryptography==40.0.2
Deprecated==1.2.13
dnspython==2.3.0
ecdsa==0.17.0
email-validator==1.3.1
fastapi==0.95.1
greenlet==2.0.2
h11==0.13.0
httptools==0.4.0
idna==3.3
install==1.3.5
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==21.3
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.9.1
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.20.0
python-multipart==0.0.6
redis==4.5.1
rsa==4.8
sentry-sdk==1.19.1
Shapely==1.8.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==2.0.5.post1
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.15
uvicorn==0.17.6
validate-email==1.3
watchgod==0.8.2
websockets==10.3
wrapt==1.14.1
qrcode
